requirements.txt
aniso8601==9.0.1
attrs==23.1.0
blinker==1.6.2
click==8.1.6
Flask==2.3.2
flask-expects-json==1.7.0
Flask-RESTful==0.3.10
flask-sqlalchemy==3.0.5
greenlet==2.0.2
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
psycopg2-binary==2.9.6
PyJWT==1.7.1
python-dotenv==1.0.0
pytz==2023.3
referencing==0.30.1
rpds-py==0.9.2
six==1.16.0
SQLAlchemy==2.0.19
typing-extensions==4.7.1
Werkzeug==2.3.6
zipp==3.16.2
